Move functions into `lib.rs`
